+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.125.0] - 2019-04-10
+### Added
+- Chronos submission
+- Preprint withdrawals
+
 ## [0.124.2] - 2019-04-08
 ### Fixed
 - make sure there are no line breaks or extra spaces in author list